refactor(plugin-multi-portal): rename portal interceptor
